integration/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
web3==6.20.0
pydantic==2.7.0
